Mejorada la clase de pruebas
